% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cost.Bundle.R
\name{Cost.Bundle}
\alias{Cost.Bundle}
\title{Find cost F}
\usage{
Cost.Bundle(n.1, n.2, c.1, c.2, alfa, beta, FC)
}
\arguments{
\item{n.1}{- number of units of good 1}

\item{n.2}{- number of units of good 2}

\item{c.1}{good 1 parameter of production cost}

\item{c.2}{good 1 parameter of production cost}

\item{alfa}{parameter of scale economics alfa = 0 --> CRS, alfa < 0 --> IRS, alfa < 0 --> DRS}

\item{beta}{parameter of sope economics  beta = 0 --> neutral, beta > 0 complementary, beta < 0 substitution}

\item{FC}{fixed Cost of production}

\item{teta}{parameter of complementary and substitution of goods beta = 0 --> neutral, beta > 0 complementary, beta < 0 substitution}
}
\value{
The TC cost function:
}
\description{
Find cost F
}
