% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Max.Profit.PB.R, R/Zero.Profit.PB.R
\name{Max.Profit.PB}
\alias{Max.Profit.PB}
\alias{Max.Profit.PB}
\title{Find maximum profit for Pure Bundling strategy}
\usage{
Max.Profit.PB(r1.r2, pb.min.max, c.1, c.2, alfa, beta, teta, FC)

Max.Profit.PB(r1.r2, pb.min.max, c.1, c.2, alfa, beta, teta, FC)
}
\arguments{
\item{r1.r2}{NX2 reservation prices of two goods []}

\item{pb.min.max}{Vector}

\item{c.1}{good 1 parameter of production cost}

\item{c.2}{good 1 parameter of production cost}

\item{alfa}{parameter of scale economics alfa = 0 --> CRS, alfa < 0 --> IRS, alfa < 0 --> DRS}

\item{beta}{parameter of sope economics  beta = 0 --> neutral, beta > 0 complementary, beta < 0 substitution}

\item{teta}{parameter of complementary and substitution of goods beta = 0 --> neutral, beta > 0 complementary, beta < 0 substitution}

\item{FC}{fixed Cost of production}

\item{r1.r2}{NX2 reservation prices of two goods []}

\item{pb.min.max}{Vector}

\item{c.1}{good 1 parameter of production cost}

\item{c.2}{good 1 parameter of production cost}

\item{alfa}{parameter of scale economics alfa = 0 --> CRS, alfa < 0 --> IRS, alfa < 0 --> DRS}

\item{beta}{parameter of sope economics  beta = 0 --> neutral, beta > 0 complementary, beta < 0 substitution}

\item{teta}{parameter of complementary and substitution of goods beta = 0 --> neutral, beta > 0 complementary, beta < 0 substitution}

\item{FC}{fixed Cost of production}
}
\value{
max.profit

max.profit
}
\description{
Find maximum profit for Pure Bundling strategy

Find maximum profit for Pure Bundling strategy
}
