% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prices.PB.R
\name{Prices.PB}
\alias{Prices.PB}
\title{This function create wector of prices for Pure Components}
\usage{
Prices.PB(pb.min.max, step)
}
\arguments{
\item{pb.min.max}{Vector}
}
\value{
wector of prices for Pure Bundling
}
\description{
This function create wector of prices for Pure Components
}
