% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profit.PC.R
\name{Profit.PC}
\alias{Profit.PC}
\title{Profit, Consumer surplus, cost, demands for Pure Componenets strategy}
\usage{
Profit.PC(r1.r2, price.pc, c.1, c.2, alfa, beta, teta, FC)
}
\arguments{
\item{r1.r2}{N valuations of good 1 and 2  - reserversion prices r1, r2}

\item{price.pc}{A vector of monopoly prices of good 1 and 2 [price.pc <- c(p1, p2)]}

\item{c.1}{Parameter of production cost of good 1}

\item{c.2}{Parameter of production cost of good 1}

\item{alfa}{Parameter of scale economics alfa = 0 --> CRS, alfa < 0 --> IRS, alfa < 0 --> DRS}

\item{beta}{Parameter of sope economics  beta = 0 --> neutral, beta > 0 complementary in proiduction, beta < 0 substitution in production}

\item{teta}{Parameter of complementary and substitution of goods beta = 0 --> neutral, beta > 0 complementary in consumption, beta < 0 substitution in consumption}

\item{FC}{Fixed Cost of production}
}
\value{
For a given: i. N valuations of r1 ,r2 ii. Monopoly prices p1, p2 iii. Technology
 profit  - profit for Pure Componenets strategy
 c.s     - Consumer surplus for Pure Componenets strategy
 t.c     - cost of production
 no.buy  - number of consumers which do not buy any good
 buy.1   - number of consumers which buy good 1
 buy.2   - number of consumers which buy good 2
 buy.1.2 - number of consumers which buy separately good 1 and good 2
}
\description{
Profit, Consumer surplus, cost, demands for Pure Componenets strategy
}
