% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r.1.r2.cable}
\alias{r.1.r2.cable}
\title{Reservation prices of TV and Interent}
\format{A data frame with columns:
\describe{
 \item{TV}{Reservation prices of TV cable services}
 \item{INT}{Reservation prices of Interet cable services}
}}
\source{
Data from reserch \url{http://www.sciencedirect.com/science/article/pii/S0308596117300435}
}
\usage{
r.1.r2.cable
}
\description{
Reservation prices of TV and Interent
}
\examples{
\dontrun{
 print(r.1.r2.cable)
}
}
\keyword{datasets}
