% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prices.MB.R
\name{Prices.MB}
\alias{Prices.MB}
\title{This function create wector of prices for Mixt Bundling}
\usage{
Prices.MB(p.1.min.max, p.2.min.max, p.b.min.max, step)
}
\arguments{
\item{pb.min.max}{Vector}
}
\value{
wector of prices for Mixt Bundling
}
\description{
This function create wector of prices for Mixt Bundling
}
