% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prices.PC.R
\name{Prices.PC}
\alias{Prices.PC}
\title{This function create wector of prices for Pure Components}
\usage{
Prices.PC(p1.min.max, p2.min.max, step)
}
\arguments{
\item{p.1.min.max}{Vector}

\item{p.2.min.max}{dfg}
}
\value{
wector of prices for Pure Components
}
\description{
This function create wector of prices for Pure Components
}
