% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profit.PB.R
\name{Profit.PB}
\alias{Profit.PB}
\title{Profit function for PB}
\usage{
Profit.PB(r1.r2, prices.pb, c.1, c.2, alfa, beta, teta, FC)
}
\arguments{
\item{r1.r2}{NX2 reservation prices of two goods []}

\item{prices.pb}{- matrix of prices}

\item{c.1}{good 1 parameter of production cost}

\item{c.2}{good 1 parameter of production cost}

\item{alfa}{parameter of scale economics alfa = 0 --> CRS, alfa < 0 --> IRS, alfa < 0 --> DRS}

\item{beta}{parameter of sope economics  beta = 0 --> neutral, beta > 0 complementary, beta < 0 substitution}

\item{teta}{parameter of complementary and substitution of goods beta = 0 --> neutral, beta > 0 complementary, beta < 0 substitution}

\item{FC}{fixed Cost of production}
}
\value{
max profit
}
\description{
Profit function for PB
}
